refactor: модульная архитектура и поддержка Streamable HTTP
- Устранены дублирования кода путем объединения server-sse.ts и index.ts
- Создана модульная архитектура с разделением ответственности:
  * src/types.ts - интерфейсы и типы данных
  * src/config.ts - конфигурация, API ключи, определения инструментов
  * src/handlers.ts - обработчики API (geocoding, maps rendering)
  * src/server.ts - фабрика MCP сервера
  * src/http-transport.ts - HTTP транспорт с поддержкой всех протоколов
  * index.ts - главная функция запуска (сокращена до 36 строк)

- Реализована полная поддержка Streamable HTTP:
  * GET/POST /mcp - новый рекомендованный endpoint
  * Корректная обработка JSON-RPC сообщений
  * Поддержка initialize, tools/list, tools/call методов

- Сохранена обратная совместимость:
  * GET /sse - legacy SSE endpoint
  * POST /message - legacy message endpoint
  * Гибридный режим работы с существующими клиентами

- Улучшена структура проекта:
  * Четкое разделение логики по модулям
  * Упрощенное тестирование и расширение
  * Правильная типизация с MCP SDK

Все endpoint'ы работают корректно в MCP Inspector.
