+++ package.json
@@ -20,7 +20,7 @@
     "prepare": "npm run build",
     "watch": "tsc --watch",
     "start": "node dist/index.js",
-    "start:sse": "node dist/server-sse.js"
+    "start:http": "node dist/index.js --transport http"
   },
   "dependencies": {
     "@modelcontextprotocol/sdk": "1.0.1",
